ailand-backend/requirements.txt
# FastAPI and Server
fastapi==0.121.2
uvicorn[standard]==0.38.0

#  Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.19
bcrypt==4.0.1

#  Validation 
pydantic==2.12.4
pydantic-settings==2.7.1
email-validator==2.2.0

#  Utilities 
python-dotenv==1.2.1

#  Database 
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

#  Testing 
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

#  Code Quality 
black==24.10.0
flake8==7.1.1
mypy==1.14.0
isort==5.13.2


#  Knowledge Graph
SPARQLWrapper==2.0.0
